requirements.txt
pywebview
llama-cpp-python
python-dotenv
pydantic
PyQt5
PyQtWebEngine
qtpy
